created a map lookup table from tilesheet
-all delta time is now a duouble rather than a float
-created a lookup tile table for the map
-rendered a map to teh screen according top the array of numbers in Map.h

+++ RPG-Game/Player.h
@@ -10,9 +10,8 @@ private:
 
 	std::vector<Bullet> bullets;
 	
-
 	float maxFireRate;
-	float fireRateTimer;
+	double fireRateTimer;
 
 	float playerSpeed;
 
@@ -26,8 +25,8 @@ public:
 	Player();
 	~Player();
 
-	void Initialize(); // <- called once app start
-	void Load(); // <-  called once per app start
-	void Update(float deltaTime, Skeleton& skeleton, sf::Vector2f& mousePosition); // <- called once per frame
-	void Draw(sf::RenderWindow& window); // <- called once per frame
+	void Initialize(); 
+	void Load(); 
+	void Update(double deltaTime, Skeleton& skeleton, sf::Vector2f& mousePosition); 
+	void Draw(sf::RenderWindow& window);
 };
